[README] And readme for all this stuff.
This commit is recommended for use until the next such
announcement. What can this code do:
  i)  Assemble an opt-code file to a binary file
  ii) Run the binary file on the machine
The assembler now supports only limited variations of argument,
namely
  i)  Two hex bytes in the following format:
       0xPQ 0xYZ
        ^    ^
        |    |
        |    L
        |     |
 (lower byte) (higher byte)
      where P, Q, Y, and Z are digits in range [0,f]; a=10, b=11,
      c=12, d=13, e=14, f=15. For byte, which are less, than 16
      (0x10), you must insert the leading zero: ie 0x0f for 15.
  ii) Label, which LASM converts to the absolute address.
What is nice to have is a conversion from ASCII chars to integers
and support not only hex constants, but also decimals and binary
arguments as well. Also, the big issue is error messages.